requirements.txt
aiofile==3.8.7
aiofiles==23.1.0
aiohttp==3.8.5
av==10.0.0
EbookLib==0.18
google-images-search==1.4.6
numpy==1.24.4
opencv-python-headless==4.8.0.74
pillow==10.0.0
piper-tts==1.2.0
pyrogram==2.0.106
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0
selenium==4.11.2
shutup==0.2.0
speedtest-cli==2.1.3
termcolor==1.1
tgcrypto==1.2.5
tinydb==4.8.0
translate==3.6.1
tronpy==0.4.0
web3==6.8.0
youtube-search-python==1.6.6
yt-dlp==2023.7.6